services/api-py/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
psycopg2-binary>=2.9.10
httpx>=0.28.0
pydantic>=2.10.0
